requirements.txt

matplotlib~=3.2.2
numpy~=1.18.5
scipy~=1.5.0
pandas~=1.0.5
seaborn~=0.10.1
scikit-learn~=0.23.1
dynamic-yaml
h5py
tqdm~=4.47.0
pytorch-crf
nltk~=3.5
fastnlp~=0.5.5
openpyxl~=3.0.4
six~=1.15.0
pytest~=5.4.3
spacy~=2.3.2
sentencepiece~=0.1.91
xlrd
flask~=1.1.2
regex~=2020.6.8
xlwt~=1.3.0
pyyaml~=5.3.1
transformers~=3.4.0
allennlp~=1.2.2
requests~=2.24.0
pyrouge~=0.1.3